+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 # CHANGELOG
 
 <p>
+<b>Version 1.0.56</b> August 2022<br/>
+- NEW: added string conversion from "true", "false", "0", "1", "close" to boolean.</br>
+- Updated the README</br>
+</p>
+<p>
 <b>Version 1.0.55</b> July 2022<br/>
 - NEW: you can now specify the input property name from witch the node picks up the payload.</br>
 - NEW: added more Homeassistant string compatibility values.</br>